refactor(vscode): replace workspace-scoped-container with workspace-scoped for improved dependency management
